+++ Challenge.Persistence/Entities/Balance.cs
@@ -5,8 +5,8 @@ namespace Challenge.Persistence.Entities
     public class Balance : BaseEntity
     {
         public Guid UserId { get; set; }
-        public long? AvailableBalance { get; set; }
-        public long? BlockedBalance { get; set; }
+        public double? AvailableBalance { get; set; }
+        public double? BlockedBalance { get; set; }
         public string? Currency { get; set; }
         public DateTime? LastUpdated { get; set; }
         public User User { get; set; }
